# openapi.yaml
openapi: 3.1.0
info:
  title: Total Recall Memory API
  description: Store and search memories with semantic search. Used by AI assistants to maintain long-term memory.
  version: 1.0.0
servers:
  - url: https://recall.stakewatch.dev
paths:
  /api/search:
    post:
      operationId: searchMemories
      summary: Search memories using semantic and keyword search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Search query text
                limit:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 50
                namespaces:
                  type: array
                  items:
                    type: string
                  description: Filter to specific namespaces
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Memory"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/store:
    post:
      operationId: storeMemory
      summary: Store a new memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, namespace]
              properties:
                content:
                  type: string
                  description: Memory content text
                namespace:
                  type: string
                  description: "Namespace to store in (e.g. work, shared)"
                tags:
                  type: array
                  items:
                    type: string
                source:
                  type: string
                  description: Source identifier (defaults to key name)
                idempotency_key:
                  type: string
                  minLength: 1
                  maxLength: 512
                  description: Optional retry key scoped only to the authenticated API key. Reusing it updates the same memory while preserving created_at; callers authorized for both namespaces may move it and may update its access level.
      responses:
        "200":
          description: Memory stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  created:
                    type: boolean
                  idempotency_key_honored:
                    type: boolean
                    description: Present and true only when an idempotency_key was supplied and honored. Keyed callers should require this acknowledgement before checkpointing.
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/store-document:
    post:
      operationId: storeDocument
      summary: Store a full document (auto-chunked for search)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, namespace]
              properties:
                title:
                  type: string
                content:
                  type: string
                  description: Nonblank document content. The authoritative limit is 1 MiB of decoded UTF-8 (not characters); content is preserved losslessly in embedding chunks of at most 2,000 UTF-8 bytes.
                namespace:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                source:
                  type: string
                idempotency_key:
                  type: string
                  description: Optional retry key scoped to the authenticated API key. Reusing it with a different document returns 409.
      responses:
        "200":
          description: Document stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  chunks:
                    type: integer
        "400":
          description: Invalid request, including blank content or decoded UTF-8 content over 1 MiB
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Idempotency conflict or incomplete prior document write
        "413":
          description: JSON transport envelope exceeds the independently configured HTTP parser limit
  /api/media/events:
    post:
      operationId: ingestMediaEvents
      summary: Ingest media playback or viewing events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  maxItems: 500
                  items:
                    type: object
                    required: [service, event_type, title, played_at]
                    properties:
                      service:
                        type: string
                      service_id:
                        type: string
                      event_type:
                        type: string
                      title:
                        type: string
                      artist:
                        type: string
                      album:
                        type: string
                      show:
                        type: string
                      season:
                        type: integer
                      episode:
                        type: integer
                      year:
                        type: integer
                      genres:
                        type: array
                        items:
                          type: string
                      duration_ms:
                        type: integer
                      played_ms:
                        type: integer
                      completed:
                        type: boolean
                      played_at:
                        type: string
                        format: date-time
                      metadata:
                        type: object
                        additionalProperties: true
      responses:
        "200":
          description: Media events ingested
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
components:
  schemas:
    Memory:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        namespace:
          type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
        final_score:
          type: number
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []